backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.0.0,<3.0.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
web3==6.11.0
requests==2.31.0
httpx==0.25.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
fabric-sdk-py==1.0.0
cryptography==41.0.7
python-multipart==0.0.6

